+++ web/src/routes/_user/household/$householdId/categories/-components/new-category.tsx
@@ -33,7 +33,7 @@ import {
   ComboboxItem,
   ComboboxList,
 } from '@/components/ui/combobox'
-import { commitMutationResult } from '@/relay'
+import { commitMutationResult } from '@/lib/relay'
 import { IconPicker, type IconName } from '@/components/ui/icon-picker'
 import { useHousehold } from '@/hooks/use-household'
